pku-pkg/requirements.txt
kubernetes==31.0.0
PyYAML==6.0.2
PyYAML==6.0.2
redis==5.0.8
